+++ src/helpers/border.js
@@ -5,37 +5,92 @@ import {
 	RangeControl,
 	SelectControl,
 	ColorPicker,
+	__experimentalHeading as Heading,
+	Button,
 } from "@wordpress/components";
-import { camelCase } from "lodash";
+import { camelCase, capitalize } from "lodash";
+
+const BORDER_STYLES = [
+	{ label: __("-- SELECT --", "beer-blocks"), value: "" },
+	{ label: __("None", "beer-blocks"), value: "none" },
+	{ label: __("Dotted", "beer-blocks"), value: "dotted" },
+	{ label: __("Dashed", "beer-blocks"), value: "dashed" },
+	{ label: __("Solid", "beer-blocks"), value: "solid" },
+	{ label: __("Double", "beer-blocks"), value: "double" },
+	{ label: __("Inset", "beer-blocks"), value: "inset" },
+	{ label: __("Outset", "beer-blocks"), value: "outset" },
+];
+
+const label = (type, side) =>
+	({
+		style: {
+			all: __("Border style", "beer-blocks"),
+			top: __("Border top style", "beer-blocks"),
+			right: __("Border right style", "beer-blocks"),
+			bottom: __("Border bottom style", "beer-blocks"),
+			left: __("Border left style", "beer-blocks"),
+		},
+		width: {
+			all: __("Border width", "beer-blocks"),
+			top: __("Border top width", "beer-blocks"),
+			right: __("Border right width", "beer-blocks"),
+			bottom: __("Border bottom width", "beer-blocks"),
+			left: __("Border left width", "beer-blocks"),
+		},
+		color: {
+			all: __("Border color", "beer-blocks"),
+			top: __("Border top color", "beer-blocks"),
+			right: __("Border right color", "beer-blocks"),
+			bottom: __("Border bottom color", "beer-blocks"),
+			left: __("Border left color", "beer-blocks"),
+		},
+	}[type][side === "" ? "all" : side]);
+
+const sectionTitle = (side) =>
+	({
+		all: (
+			<Heading align="center" level="3">
+				{__("ALL BORDERS", "beer-blocks")}
+			</Heading>
+		),
+		top: (
+			<Heading align="center" level="3">
+				{__("BORDER TOP", "beer-blocks")}
+			</Heading>
+		),
+		right: (
+			<Heading align="center" level="3">
+				{__("BORDER RIGHT", "beer-blocks")}
+			</Heading>
+		),
+		bottom: (
+			<Heading align="center" level="3">
+				{__("BORDER BOTTOM", "beer-blocks")}
+			</Heading>
+		),
+		left: (
+			<Heading align="center" level="3">
+				{__("BORDER LEFT", "beer-blocks")}
+			</Heading>
+		),
+	}[side === "" ? "all" : side]);
 
 export const borderStyleAttribute = () => ({ type: "string", default: "" });
 
-export const borderStyleControl = (
+export const borderStyleControl = ({
 	props,
 	attrName = "borderStyle",
-	label = __("Style", "beer-blocks")
-) => {
+	label = __("Border style", "beer-blocks"),
+}) => {
 	const {
 		setAttributes,
-		attributes: { [attrName]: borderStyle },
+		attributes: { [attrName]: borderStyle = undefined },
 	} = props;
 
-	return (
+	return borderStyle !== undefined ? (
 		<SelectControl
 			label={label}
-			options={[
-				{ label: __("-- SELECT --", "beer-blocks"), value: "" },
-				{ label: "None", value: "none" },
-				{ label: "Dotted", value: "dotted" },
-				{ label: "Dashed", value: "dashed" },
-				{ label: "Solid", value: "solid" },
-				{ label: "Double", value: "double" },
-				{ label: "Groove", value: "groove" },
-				{ label: "Ridge", value: "ridge" },
-				{ label: "Inset", value: "inset" },
-				{ label: "Outset", value: "outset" },
-				{ label: "Hidden", value: "hidden" },
-			]}
+			options={BORDER_STYLES}
 			value={borderStyle}
 			onChange={(value) =>
 				setAttributes({
@@ -43,107 +98,127 @@ export const borderStyleControl = (
 				})
 			}
 		/>
-	);
+	) : null;
 };
 
-export const borderStyleStyles = (borderStyle) =>
-	borderStyle ? { borderStyle } : {};
+export const borderStyleStyles = (borderStyle, side = "") =>
+	borderStyle ? { [`border${capitalize(side)}Style`]: borderStyle } : {};
 
-export const borderWidthAttribute = () => ({ type: "string", default: "" });
+export const borderWidthAttribute = () => ({
+	type: "string",
+	default: "",
+});
 
-export const borderWidthControl = (
+export const borderWidthControl = ({
 	props,
 	attrName = "borderWidth",
-	label = __("Width", "beer-blocks")
-) => {
+	label = __("Border width", "beer-blocks"),
+}) => {
 	const {
 		setAttributes,
-		attributes: { [attrName]: borderWidth },
+		attributes: { [attrName]: borderWidth = undefined },
 	} = props;
 
-	return (
+	return borderWidth !== undefined ? (
 		<BaseControl>
 			<RangeControl
 				label={label}
-				value={borderWidth}
-				onChange={(value) => setAttributes({ [attrName]: value })}
+				value={borderWidth ? parseInt(borderWidth.replace(/px$/, "")) : ""}
+				onChange={(value) =>
+					setAttributes({ [attrName]: value ? `${value}px` : "" })
+				}
 				min={1}
 				step={1}
 				allowReset
+				resetFallbackValue={""}
 			/>
 		</BaseControl>
-	);
+	) : null;
 };
 
-export const borderWidthStyles = (borderWidth) =>
-	borderWidth ? { borderWidth } : {};
+export const borderWidthStyles = (borderWidth, side = "") =>
+	borderWidth ? { [`border${capitalize(side)}Width`]: borderWidth } : {};
 
 export const borderColorAttribute = () => ({ type: "string", default: "" });
 
-export const borderColorControl = (
+export const borderColorControl = ({
 	props,
 	attrName = "borderColor",
-	label = __("Color", "beer-blocks")
-) => {
+	label = __("Border color", "beer-blocks"),
+}) => {
 	const {
 		setAttributes,
-		attributes: { [attrName]: borderColor },
+		attributes: { [attrName]: borderColor = undefined },
 	} = props;
 
-	return (
+	return borderColor !== undefined ? (
 		<BaseControl label={label}>
 			<ColorPicker
 				color={borderColor}
 				onChangeComplete={(value) => setAttributes({ [attrName]: value.hex })}
 				disableAlpha
+				defaultValue=""
 			/>
 		</BaseControl>
-	);
+	) : null;
 };
 
-export const borderColorStyles = (borderColor) =>
-	borderColor ? { borderColor } : {};
+export const borderColorStyles = (borderColor, side = "") =>
+	borderColor ? { [`border${capitalize(side)}Color`]: borderColor } : {};
 
-export const attributes = (attrPrefixName = "") => ({
-	[camelCase(`${attrPrefixName}-border-style`)]: {
+export const attributes = ({ attrPrefixName = "", side = "" } = {}) => ({
+	[camelCase(`${attrPrefixName}-border-${side}-style`)]: {
 		type: "string",
 		default: "",
 	},
-	[camelCase(`${attrPrefixName}-border-width`)]: {
+	[camelCase(`${attrPrefixName}-border-${side}-width`)]: {
 		type: "string",
 		default: "",
 	},
-	[camelCase(`${attrPrefixName}-border-color`)]: {
+	[camelCase(`${attrPrefixName}-border-${side}-color`)]: {
 		type: "string",
 		default: "",
 	},
 });
 
-export const innerControls = (
+export const innerControls = ({
 	props,
 	attrPrefixName = "",
-	labels = {
-		style: __("Style", "beer-blocks"),
-		width: __("Width", "beer-blocks"),
-		color: __("Color", "beer-blocks"),
-	}
-) => (
+	side = "",
+	title = undefined,
+}) => (
 	<>
-		{borderStyleControl(
+		{title === undefined ? sectionTitle(side) : title}
+		{borderStyleControl({
 			props,
-			camelCase(`${attrPrefixName}-border-style`),
-			labels.style
-		)}
-		{borderWidthControl(
+			attrName: camelCase(`${attrPrefixName}-border-${side}-style`),
+			label: label("style", side),
+		})}
+		{borderWidthControl({
 			props,
-			camelCase(`${attrPrefixName}-border-width`),
-			labels.width
-		)}
-		{borderColorControl(
+			attrName: camelCase(`${attrPrefixName}-border-${side}-width`),
+			label: label("width", side),
+		})}
+		{borderColorControl({
 			props,
-			camelCase(`${attrPrefixName}-border-color`),
-			labels.color
-		)}
+			attrName: camelCase(`${attrPrefixName}-border-${side}-color`),
+			label: label("color", side),
+		})}
+		<BaseControl>
+			<Button
+				className="is-destructive"
+				style={{ display: "block", textAlign: "center", width: "100%" }}
+				onClick={() =>
+					props.setAttributes({
+						[camelCase(`${attrPrefixName}-border-${side}-style`)]: "",
+						[camelCase(`${attrPrefixName}-border-${side}-width`)]: "",
+						[camelCase(`${attrPrefixName}-border-${side}-color`)]: "",
+					})
+				}
+			>
+				{__("Clear fields", "beer-blocks")}
+			</Button>
+		</BaseControl>
 	</>
 );
 
@@ -151,23 +226,97 @@ export const controls = ({
 	props,
 	initialOpen = false,
 	attrPrefixName = "",
-}) => (
-	<PanelBody title={__("Border", "beer-blocks")} initialOpen={initialOpen}>
-		{innerControls(props, attrPrefixName)}
-	</PanelBody>
-);
+}) => {
+	const {
+		[camelCase(`${attrPrefixName}-border-style`)]: borderStyle = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-top-style`
+		)]: borderTopStyle = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-right-style`
+		)]: borderRightStyle = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-bottom-style`
+		)]: borderBottomStyle = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-left-style`
+		)]: borderLeftStyle = undefined,
+	} = props.attributes;
+
+	return (
+		<PanelBody title={__("Border", "beer-blocks")} initialOpen={initialOpen}>
+			{borderStyle !== undefined && innerControls({ props, attrPrefixName })}
+			{borderTopStyle !== undefined &&
+				innerControls({ props, attrPrefixName, side: "top" })}
+			{borderRightStyle !== undefined &&
+				innerControls({ props, attrPrefixName, side: "right" })}
+			{borderBottomStyle !== undefined &&
+				innerControls({ props, attrPrefixName, side: "bottom" })}
+			{borderLeftStyle !== undefined &&
+				innerControls({ props, attrPrefixName, side: "left" })}
+		</PanelBody>
+	);
+};
 
 export const styles = (attributes, attrPrefixName = "") => {
 	const {
-		[camelCase(`${attrPrefixName}-border-style`)]: borderStyle,
-		[camelCase(`${attrPrefixName}-border-width`)]: borderWidth,
-		[camelCase(`${attrPrefixName}-border-color`)]: borderColor,
+		[camelCase(`${attrPrefixName}-border-style`)]: borderStyle = undefined,
+		[camelCase(`${attrPrefixName}-border-width`)]: borderWidth = undefined,
+		[camelCase(`${attrPrefixName}-border-color`)]: borderColor = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-top-style`
+		)]: borderTopStyle = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-top-width`
+		)]: borderTopWidth = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-top-color`
+		)]: borderTopColor = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-right-style`
+		)]: borderRightStyle = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-right-width`
+		)]: borderRightWidth = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-right-color`
+		)]: borderRightColor = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-bottom-style`
+		)]: borderBottomStyle = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-bottom-width`
+		)]: borderBottomWidth = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-bottom-color`
+		)]: borderBottomColor = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-left-style`
+		)]: borderLeftStyle = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-left-width`
+		)]: borderLeftWidth = undefined,
+		[camelCase(
+			`${attrPrefixName}-border-left-color`
+		)]: borderLeftColor = undefined,
 	} = attributes;
 
 	return {
 		...borderStyleStyles(borderStyle),
+		...borderStyleStyles(borderTopStyle, "top"),
+		...borderStyleStyles(borderRightStyle, "right"),
+		...borderStyleStyles(borderBottomStyle, "bottom"),
+		...borderStyleStyles(borderLeftStyle, "left"),
 		...borderWidthStyles(borderWidth),
+		...borderWidthStyles(borderTopWidth, "top"),
+		...borderWidthStyles(borderRightWidth, "right"),
+		...borderWidthStyles(borderBottomWidth, "bottom"),
+		...borderWidthStyles(borderLeftWidth, "left"),
 		...borderColorStyles(borderColor),
+		...borderColorStyles(borderTopColor, "top"),
+		...borderColorStyles(borderRightColor, "right"),
+		...borderColorStyles(borderBottomColor, "bottom"),
+		...borderColorStyles(borderLeftColor, "left"),
 	};
 };